Fix race condition gopool

+++ test/integration/bundle/bundle_test.go
@@ -572,5 +572,5 @@ func TestDeeplyNestedInvocation(t *testing.T) {
 	assert.True(t, wfi.Status.Successful())
 
 	output := typedvalues.MustFormat(wfi.Status.Output)
-	assert.Equal(t, 5, output)
+	assert.Equal(t, float64(5), output)
 }
